requirements.txt
# Core Dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# AI/ML Dependencies
torch==2.1.2
torchvision==0.16.2
transformers==4.35.2
faiss-cpu==1.7.4
numpy==1.26.4
Pillow==10.2.0
scikit-learn==1.4.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.8.6
pandas==2.2.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiofiles==23.2.1

# Added from the code block
streamlit==1.31.1
Pillow==10.2.0
python-dotenv==1.0.1 